move processor benchmarks to processor_timing_test.go for consistency with other benchmarks
